refactoring(2), 시나리오테스트 통과
