+++ packages/assignment-5/src/origin/components/CartPage.tsx
@@ -10,18 +10,19 @@ export const CartPage = ({ products, coupons }: Props) => {
   const [cart, setCart] = useState<CartItem[]>([]);
   const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
 
-
   const addToCart = (product: Product) => {
     const remainingStock = getRemainingStock(product);
     if (remainingStock <= 0) return;
 
-    setCart(prevCart => {
-      const existingItem = prevCart.find(item => item.product.id === product.id);
+    setCart((prevCart) => {
+      const existingItem = prevCart.find(
+        (item) => item.product.id === product.id,
+      );
       if (existingItem) {
-        return prevCart.map(item =>
+        return prevCart.map((item) =>
           item.product.id === product.id
             ? { ...item, quantity: Math.min(item.quantity + 1, product.stock) }
-            : item
+            : item,
         );
       }
       return [...prevCart, { product, quantity: 1 }];
@@ -29,19 +30,28 @@ export const CartPage = ({ products, coupons }: Props) => {
   };
 
   const removeFromCart = (productId: string) => {
-    setCart(prevCart => prevCart.filter(item => item.product.id !== productId));
+    setCart((prevCart) =>
+      prevCart.filter((item) => item.product.id !== productId),
+    );
   };
 
   const updateQuantity = (productId: string, newQuantity: number) => {
-    setCart(prevCart =>
-      prevCart.map(item => {
-        if (item.product.id === productId) {
-          const maxQuantity = item.product.stock;
-          const updatedQuantity = Math.max(0, Math.min(newQuantity, maxQuantity));
-          return updatedQuantity > 0 ? { ...item, quantity: updatedQuantity } : null;
-        }
-        return item;
-      }).filter((item): item is CartItem => item !== null)
+    setCart((prevCart) =>
+      prevCart
+        .map((item) => {
+          if (item.product.id === productId) {
+            const maxQuantity = item.product.stock;
+            const updatedQuantity = Math.max(
+              0,
+              Math.min(newQuantity, maxQuantity),
+            );
+            return updatedQuantity > 0
+              ? { ...item, quantity: updatedQuantity }
+              : null;
+          }
+          return item;
+        })
+        .filter((item): item is CartItem => item !== null),
     );
   };
 
@@ -49,13 +59,15 @@ export const CartPage = ({ products, coupons }: Props) => {
     let totalBeforeDiscount = 0;
     let totalAfterDiscount = 0;
 
-    cart.forEach(item => {
+    cart.forEach((item) => {
       const { price } = item.product;
       const { quantity } = item;
       totalBeforeDiscount += price * quantity;
 
       const discount = item.product.discounts.reduce((maxDiscount, d) => {
-        return quantity >= d.quantity && d.rate > maxDiscount ? d.rate : maxDiscount;
+        return quantity >= d.quantity && d.rate > maxDiscount
+          ? d.rate
+          : maxDiscount;
       }, 0);
 
       totalAfterDiscount += price * quantity * (1 - discount);
@@ -66,9 +78,12 @@ export const CartPage = ({ products, coupons }: Props) => {
     // 쿠폰 적용
     if (selectedCoupon) {
       if (selectedCoupon.discountType === 'amount') {
-        totalAfterDiscount = Math.max(0, totalAfterDiscount - selectedCoupon.discountValue);
+        totalAfterDiscount = Math.max(
+          0,
+          totalAfterDiscount - selectedCoupon.discountValue,
+        );
       } else {
-        totalAfterDiscount *= (1 - selectedCoupon.discountValue / 100);
+        totalAfterDiscount *= 1 - selectedCoupon.discountValue / 100;
       }
       totalDiscount = totalBeforeDiscount - totalAfterDiscount;
     }
@@ -76,21 +91,21 @@ export const CartPage = ({ products, coupons }: Props) => {
     return {
       totalBeforeDiscount: Math.round(totalBeforeDiscount),
       totalAfterDiscount: Math.round(totalAfterDiscount),
-      totalDiscount: Math.round(totalDiscount)
+      totalDiscount: Math.round(totalDiscount),
     };
   };
-
 
   const getMaxDiscount = (discounts: { quantity: number; rate: number }[]) => {
     return discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
   };
 
   const getRemainingStock = (product: Product) => {
-    const cartItem = cart.find(item => item.product.id === product.id);
+    const cartItem = cart.find((item) => item.product.id === product.id);
     return product.stock - (cartItem?.quantity || 0);
   };
 
-  const { totalBeforeDiscount, totalAfterDiscount, totalDiscount } = calculateTotal()
+  const { totalBeforeDiscount, totalAfterDiscount, totalDiscount } =
+    calculateTotal();
 
   const getAppliedDiscount = (item: CartItem) => {
     const { discounts } = item.product;
@@ -115,21 +130,31 @@ export const CartPage = ({ products, coupons }: Props) => {
         <div>
           <h2 className="text-2xl font-semibold mb-4">상품 목록</h2>
           <div className="space-y-2">
-            {products.map(product => {
+            {products.map((product) => {
               const remainingStock = getRemainingStock(product);
               return (
-                <div key={product.id} data-testid={`product-${product.id}`} className="bg-white p-3 rounded shadow">
+                <div
+                  key={product.id}
+                  data-testid={`product-${product.id}`}
+                  className="bg-white p-3 rounded shadow"
+                >
                   <div className="flex justify-between items-center mb-2">
                     <span className="font-semibold">{product.name}</span>
-                    <span className="text-gray-600">{product.price.toLocaleString()}원</span>
+                    <span className="text-gray-600">
+                      {product.price.toLocaleString()}원
+                    </span>
                   </div>
                   <div className="text-sm text-gray-500 mb-2">
-                    <span className={`font-medium ${remainingStock > 0 ? 'text-green-600' : 'text-red-600'}`}>
+                    <span
+                      className={`font-medium ${remainingStock > 0 ? 'text-green-600' : 'text-red-600'}`}
+                    >
                       재고: {remainingStock}개
                     </span>
                     {product.discounts.length > 0 && (
                       <span className="ml-2 font-medium text-blue-600">
-                        최대 {(getMaxDiscount(product.discounts) * 100).toFixed(0)}% 할인
+                        최대{' '}
+                        {(getMaxDiscount(product.discounts) * 100).toFixed(0)}%
+                        할인
                       </span>
                     )}
                   </div>
@@ -137,7 +162,8 @@ export const CartPage = ({ products, coupons }: Props) => {
                     <ul className="list-disc list-inside text-sm text-gray-500 mb-2">
                       {product.discounts.map((discount, index) => (
                         <li key={index}>
-                          {discount.quantity}개 이상: {(discount.rate * 100).toFixed(0)}% 할인
+                          {discount.quantity}개 이상:{' '}
+                          {(discount.rate * 100).toFixed(0)}% 할인
                         </li>
                       ))}
                     </ul>
@@ -162,31 +188,38 @@ export const CartPage = ({ products, coupons }: Props) => {
           <h2 className="text-2xl font-semibold mb-4">장바구니 내역</h2>
 
           <div className="space-y-2">
-            {cart.map(item => {
+            {cart.map((item) => {
               const appliedDiscount = getAppliedDiscount(item);
               return (
-                <div key={item.product.id} className="flex justify-between items-center bg-white p-3 rounded shadow">
+                <div
+                  key={item.product.id}
+                  className="flex justify-between items-center bg-white p-3 rounded shadow"
+                >
                   <div>
                     <span className="font-semibold">{item.product.name}</span>
-                    <br/>
+                    <br />
                     <span className="text-sm text-gray-600">
-                  {item.product.price}원 x {item.quantity}
+                      {item.product.price}원 x {item.quantity}
                       {appliedDiscount > 0 && (
                         <span className="text-green-600 ml-1">
-                      ({(appliedDiscount * 100).toFixed(0)}% 할인 적용)
-                    </span>
+                          ({(appliedDiscount * 100).toFixed(0)}% 할인 적용)
+                        </span>
                       )}
-                </span>
+                    </span>
                   </div>
                   <div>
                     <button
-                      onClick={() => updateQuantity(item.product.id, item.quantity - 1)}
+                      onClick={() =>
+                        updateQuantity(item.product.id, item.quantity - 1)
+                      }
                       className="bg-gray-300 text-gray-800 px-2 py-1 rounded mr-1 hover:bg-gray-400"
                     >
                       -
                     </button>
                     <button
-                      onClick={() => updateQuantity(item.product.id, item.quantity + 1)}
+                      onClick={() =>
+                        updateQuantity(item.product.id, item.quantity + 1)
+                      }
                       className="bg-gray-300 text-gray-800 px-2 py-1 rounded mr-1 hover:bg-gray-400"
                     >
                       +
@@ -212,14 +245,20 @@ export const CartPage = ({ products, coupons }: Props) => {
               <option value="">쿠폰 선택</option>
               {coupons.map((coupon, index) => (
                 <option key={coupon.code} value={index}>
-                  {coupon.name} - {coupon.discountType === 'amount' ? `${coupon.discountValue}원` : `${coupon.discountValue}%`}
+                  {coupon.name} -{' '}
+                  {coupon.discountType === 'amount'
+                    ? `${coupon.discountValue}원`
+                    : `${coupon.discountValue}%`}
                 </option>
               ))}
             </select>
             {selectedCoupon && (
               <p className="text-green-600">
-                적용된 쿠폰: {selectedCoupon.name}
-                ({selectedCoupon.discountType === 'amount' ? `${selectedCoupon.discountValue}원` : `${selectedCoupon.discountValue}%`} 할인)
+                적용된 쿠폰: {selectedCoupon.name}(
+                {selectedCoupon.discountType === 'amount'
+                  ? `${selectedCoupon.discountValue}원`
+                  : `${selectedCoupon.discountValue}%`}{' '}
+                할인)
               </p>
             )}
           </div>
@@ -228,7 +267,9 @@ export const CartPage = ({ products, coupons }: Props) => {
             <h2 className="text-2xl font-semibold mb-2">주문 요약</h2>
             <div className="space-y-1">
               <p>상품 금액: {totalBeforeDiscount.toLocaleString()}원</p>
-              <p className="text-green-600">할인 금액: {totalDiscount.toLocaleString()}원</p>
+              <p className="text-green-600">
+                할인 금액: {totalDiscount.toLocaleString()}원
+              </p>
               <p className="text-xl font-bold">
                 최종 결제 금액: {totalAfterDiscount.toLocaleString()}원
               </p>
