+++ packages/assignment-5/src/refactoring/hooks/utils/cartUtils.ts
@@ -81,3 +81,15 @@ export const updateCartItemQuantity = (
     })
     .filter((item) => item.quantity > 0);
 };
+
+export const getAppliedDiscount = (item: CartItem) => {
+  const { discounts } = item.product;
+  const { quantity } = item;
+  let appliedDiscount = 0;
+  for (const discount of discounts) {
+    if (quantity >= discount.quantity) {
+      appliedDiscount = Math.max(appliedDiscount, discount.rate);
+    }
+  }
+  return appliedDiscount;
+};
