+++ packages/assignment-5/src/refactoring/hooks/utils/productUtils.ts
@@ -1,0 +1,5 @@
+export const getMaxDiscount = (
+  discounts: { quantity: number; rate: number }[],
+) => {
+  return discounts.reduce((max, discount) => Math.max(max, discount.rate), 0);
+};
